+++ packages/job-components/src/builtin/delay/processor.ts
@@ -1,11 +1,10 @@
-import { promisify } from 'util';
+import { DataEntity, pDelay } from '@terascope/utils';
 import { DelayConfig } from './interfaces';
-import { BatchProcessor, DataEntity } from '../../operations';
-const delay = promisify(setTimeout);
+import { BatchProcessor } from '../../operations';
 
 export default class Delay extends BatchProcessor<DelayConfig> {
     async onBatch(data: DataEntity[]) {
-        await delay(this.opConfig.ms);
+        await pDelay(this.opConfig.ms);
         return data;
     }
 }
