+++ packages/job-components/src/builtin/noop/processor.ts
@@ -1,4 +1,5 @@
-import { BatchProcessor, DataEntity } from '../../operations';
+import { DataEntity } from '@terascope/utils';
+import { BatchProcessor } from '../../operations';
 
 export default class Noop extends BatchProcessor {
     async onBatch(data: DataEntity[]) {
