+++ packages/job-components/src/builtin/test-reader/slicer.ts
@@ -1,8 +1,8 @@
 import fs from 'fs';
 import path from 'path';
+import { parseJSON } from '@terascope/utils';
 import { TestReaderConfig } from './interfaces';
 import { Slicer } from '../../operations';
-import { parseJSON } from '../../utils';
 import { dataClone } from './utils';
 import defaultData from './data/slicer-data.json';
