+++ packages/job-components/src/interfaces/index.ts
@@ -1,5 +1,4 @@
 export * from './context';
 export * from './jobs';
-export * from './logger';
 export * from './operation-lifecycle';
 export * from './operations';
