+++ packages/job-components/src/operation-loader.ts
@@ -1,6 +1,7 @@
 /* tslint:disable:variable-name */
 import fs from 'fs';
 import path from 'path';
+import { isString, uniq, parseError, cloneDeep } from '@terascope/utils';
 import { LegacyOperation } from './interfaces';
 import {
     OperationAPIConstructor,
@@ -14,7 +15,6 @@ import {
     ReaderModule,
 } from './operations';
 import { readerShim, processorShim } from './operations/shims';
-import { isString, uniq, parseError, cloneDeep } from './utils';
 
 export interface LoaderOptions {
     /** Path to teraslice lib directory */
