+++ packages/job-components/src/operations/index.ts
@@ -1,6 +1,5 @@
 import BatchProcessor from './batch-processor';
 import ConvictSchema from './convict-schema';
-import DataEntity from './data-entity';
 import EachProcessor from './each-processor';
 import Fetcher from './fetcher';
 import FilterProcessor from './filter-processor';
@@ -17,7 +16,6 @@ export * from './core';
 export {
     BatchProcessor,
     ConvictSchema,
-    DataEntity,
     EachProcessor,
     Fetcher,
     FilterProcessor,
