+++ packages/job-components/src/test-helpers.ts
@@ -1,81 +1,18 @@
 
-import debugFn from 'debug';
-import { EventEmitter } from 'events';
 import path from 'path';
+import { EventEmitter } from 'events';
 import * as i from './interfaces';
-import { random, isString, uniq, getTypeOf, isFunction } from './utils';
-
-interface DebugParamObj {
-    module: string;
-    assignment?: string;
-    [name: string]: any;
-}
+import {
+    random,
+    isString,
+    getTypeOf,
+    isFunction,
+    debugLogger,
+    Logger
+} from '@terascope/utils';
 
 function newId(prefix: string): string {
     return `${prefix}-${random(10000, 99999)}`;
-}
-
-type debugParam = DebugParamObj | string;
-
-export function debugLogger(testName: string, param?: debugParam, otherName?: string): i.Logger {
-    const logger: i.Logger = new EventEmitter() as i.Logger;
-
-    const parts: string[] = [testName];
-    if (testName.indexOf('teraslice') < 0) {
-        parts.unshift('teraslice');
-    }
-
-    if (param) {
-        if (isString(param)) {
-            parts.push(param as string);
-        } else if (typeof param === 'object') {
-            parts.push(param.module);
-            if (param.assignment) {
-                parts.push(param.assignment);
-            }
-        }
-    }
-
-    if (otherName) {
-        parts.push(otherName);
-    }
-
-    const name = uniq(parts).join(':');
-
-    logger.streams = [];
-
-    logger.addStream = function (stream) {
-        // @ts-ignore
-        this.streams.push(stream);
-    };
-
-    logger.child = (opts: debugParam) => debugLogger(name, opts);
-    logger.flush = () => Promise.resolve();
-    logger.reopenFileStreams = () => {};
-    logger.level = () => 50;
-    // @ts-ignore
-    logger.levels = () => 50;
-
-    logger.src = false;
-
-    const levels = [
-        'trace',
-        'debug',
-        'info',
-        'warn',
-        'error',
-        'fatal'
-    ];
-
-    for (const level of levels) {
-        const fLevel = `[${level.toUpperCase()}]`;
-        const debug = debugFn(name);
-        logger[level] = (...args: any[]) => {
-            debug(fLevel, ...args);
-        };
-    }
-
-    return logger;
 }
 
 export function newTestSlice(request: i.SliceRequest = {}): i.Slice {
@@ -201,7 +138,7 @@ const _cachedClients = new WeakMap<TestContext, CachedClients>();
 const _createClientFns = new WeakMap<TestContext, ClientFactoryFns>();
 
 export class TestContext implements i.Context {
-    logger: i.Logger;
+    logger: Logger;
     sysconfig: i.SysConfig;
     cluster: i.ContextClusterConfig;
     apis: TestContextAPIs|i.WorkerContextAPIs|i.ContextAPIs;
@@ -274,7 +211,7 @@ export class TestContext implements i.Context {
 
         this.apis = {
             foundation: {
-                makeLogger(...params: any[]): i.Logger {
+                makeLogger(...params: any[]): Logger {
                     return debugLogger(testName, ...params);
                 },
                 getConnection(options: i.ConnectionConfig): { client: any } {
