+++ packages/job-components/test/operations/operation-api-spec.ts
@@ -28,9 +28,9 @@ describe('OperationAPI', () => {
     });
 
     beforeAll(() => {
-
         const exContextApi = new ExecutionContextAPI(context, exConfig);
         exContextApi.addToRegistry('example/api', ExampleOperationAPI);
+
         context.apis.registerAPI('executionContext', exContextApi);
     });
