+++ packages/utils/README.md
@@ -1,0 +1,22 @@
+# terascope-utils
+
+> A collection of Teraslice Utilities
+
+# Installation
+
+```bash
+# Using yarn
+yarn add @terascope/terascope-utils
+# Using npm
+npm install --save @terascope/terascope-utils
+```
+
+## Contributing
+
+Pull requests are welcome. For major changes, please open an issue first to discuss what you would like to change.
+
+Please make sure to update tests as appropriate.
+
+## License
+
+[MIT](./LICENSE) licensed.
