+++ packages/utils/jest.config.js
@@ -1,0 +1,3 @@
+'use strict';
+
+module.exports = require('../../jest.config.base')(__dirname);
