+++ packages/utils/tsconfig.build.json
@@ -1,0 +1,22 @@
+{
+    "extends": "./tsconfig",
+    "compilerOptions": {
+        "rootDir": "src",
+        "paths": {
+            "@terascope/*": [
+                "../*",
+                "../*/types"
+            ],
+            "*": [
+                "types/*",
+                "../*",
+                "../../types/*",
+                "node_modules/*",
+                "../../node_modules/*/types"
+            ]
+        }
+    },
+    "include": [
+        "src/**/*"
+    ]
+}
