+++ packages/utils/tsconfig.json
@@ -1,0 +1,11 @@
+{
+    "extends": "../../tsconfig",
+    "compilerOptions": {
+        "baseUrl": ".",
+        "outDir": "dist"
+    },
+    "include": [
+        "./src/**/*",
+        "./test/**/*"
+    ]
+}
